requirements.txt
hvac==2.4.0
boto3==1.40.74
pyyaml==6.0.3
kubernetes-asyncio==33.3.0
aiohttp==3.13.2
pgadmin4==9.10
textual==6.6.0
types-hvac==2.4.0.20251115
types-PyYAML==6.0.12.20250915
boto3-stubs==1.40.74
git+https://github.com/kubevirt/client-python.git@main#egg=kubevirt
